+++ LoginForm.BL/Services/VotingAlgorithmService.cs
@@ -14,17 +14,18 @@ namespace LoginForm.BL.Services
     {
         private readonly IVotingResultRepository _votingResultRepository;
         private readonly IVotingProcessor _votingProcessor;
+
         public VotingAlgorithmService(IVotingResultRepository votingResultRepository, IVotingProcessor votingProcessor)
         {
             _votingResultRepository = votingResultRepository;
             _votingProcessor = votingProcessor;
         }
 
-        public async Task<IEnumerable<VotingExecutionResult>> GetResults(int votingId)
+        public async Task<IEnumerable<VotingExecutionResult>> GetResults(long votingId)
         {
             var votingResults = await _votingResultRepository.GetByVoting(votingId);
-
             var results = _votingProcessor.Handle(votingResults);
+
             return results;
         }
     }
